examples/social/shape-mail/requirements.txt
flask
sentry-sdk[flask]
requests
python-dotenv
openai
PyPDF2